requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-import-export==3.3.6
Pillow==10.2.0
django-taggit==5.0.1
django-widget-tweaks==1.5.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7 